# design/gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"Book Service","description":"Service that provides CRUD operations for a book","version":"1.0"},"servers":[{"url":"http://localhost:8000/book","description":"book hosts the Book Service."}],"paths":{"/book":{"post":{"tags":["book"],"summary":"postBook book","operationId":"book#postBook","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostBookRequestBody"},"example":{"author":"Dolorum et repellendus vero id voluptatem.","bookCover":"UXVpZGVtIGNvbnNlcXVhdHVyIG9jY2FlY2F0aSBleGNlcHR1cmkgZWl1cyBleGVyY2l0YXRpb25lbS4=","id":7940037928223836396,"publishedAt":"Quidem cum.","title":"Ut et quo."}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostBookResponseBodyResultOperation"},"example":{"id":3602719749388078460}}}},"400":{"description":"BadRequest: Bad Request response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/book/{bookID}":{"get":{"tags":["book"],"summary":"getBook book","operationId":"book#getBook","parameters":[{"name":"bookID","in":"path","description":"Book ID","required":true,"schema":{"type":"integer","description":"Book ID","example":4575182873021229415,"format":"int64"},"example":5534387086196142826}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetBookResponseBody"},"example":{"author":"Nostrum voluptas illum qui.","bookCover":"SWxsbyBlc3Qgc2VkIHJhdGlvbmUgaXRhcXVlLg==","id":3338801228728578773,"publishedAt":"Sunt similique molestias ipsum.","title":"Laborum magnam sint voluptas qui illo non."}}}},"400":{"description":"BadRequest: Bad Request response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"NotFound: Not Found response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/swagger.json":{"get":{"tags":["book"],"summary":"Download ./gen/http/openapi.json","operationId":"book#/swagger.json","responses":{"200":{"description":"File downloaded"}}}}},"components":{"schemas":{"Error":{"type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":true}},"example":{"fault":true,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]},"GetBookResponseBody":{"type":"object","properties":{"author":{"type":"string","description":"author of the book","example":"Porro distinctio aliquid provident modi dolor."},"bookCover":{"type":"string","description":"cover of the book","example":"VmVuaWFtIHF1aWEgaGljIGlkLg==","format":"binary"},"id":{"type":"integer","description":"ID of the book","example":242785944888730846,"format":"int64"},"publishedAt":{"type":"string","description":"cover of the book","example":"Libero cupiditate voluptatem similique."},"title":{"type":"string","description":" of the book","example":"Sunt assumenda adipisci iste eius."}},"description":"GetBookResponseBody result type (default view)","example":{"author":"Voluptas earum eaque quo dolorem repellendus.","bookCover":"QWNjdXNhbnRpdW0gc2VkLg==","id":1940675957994074533,"publishedAt":"Consectetur nulla consequatur possimus et ratione.","title":"Vel hic expedita."}},"PostBookRequestBody":{"type":"object","properties":{"author":{"type":"string","description":"author of the book","example":"Quasi enim error quis quia in quasi."},"bookCover":{"type":"string","description":"cover of the book","example":"QW1ldCBkb2xvciBhdXQu","format":"binary"},"id":{"type":"integer","description":"ID of the book","example":7722079312788365161,"format":"int64"},"publishedAt":{"type":"string","description":"cover of the book","example":"Eum sed quo omnis est dolores illum."},"title":{"type":"string","description":" of the book","example":"Suscipit ipsam sequi quia et."}},"example":{"author":"Et distinctio.","bookCover":"TWFnbmkgcXVhbSBtb2RpLg==","id":6182332736197887162,"publishedAt":"Beatae placeat.","title":"Deleniti est architecto qui aut dolor."}},"PostBookResponseBodyResultOperation":{"type":"object","properties":{"id":{"type":"integer","description":"ID of the book","example":5830167137320991927,"format":"int64"}},"description":"PostBookResponseBody result type (resultOperation view)","example":{"id":3417202608605971474}}}},"tags":[{"name":"book","description":"The book service performs operations for books"}]}